Added export_monster.py, sample_monster.monster
export_monster.py is mean to take the json data for a monster and export it to a .monster file to support homebrew statblocks. Currently, we templated out what it should do but didn't actually write in the data parsing, as we are in the middle of comparing the data of the two different json formats.

sample_monster.monster is a template .monster file that can be filled in.