.json to .monster keys added--unfinished
This is a json file of the form {tag_in_monster_file: corresponding_tag_in_json_with_info} so that we can find the necessary information from the json file and put it into the monster file. Work in progress--there are a lot of variables in the .monster I haven't found a match for yet.