Progress on BBEGinator logo
I have finished the lineart around the dragon "B". Still need to do lineart for the rest of the letters, add the fire, and decide on colors. I think having a version with colors (and either a white or transparent background), a version in black with a transparent background to layer on light colors, and a version in white with a transparent background to layer on dark colors would be ideal, so I will make all 3 once lineart is complete.