Search.js payload update
Updated the payload of search.js in the following way:
The main "payload" has all of the categories that mgetter uses now (prior was missing all the stuff like slug and name).
A new "payload2" was made to store all the categories that were present in the search.js payload but are not currently used in mgetter at all.