Working call to algorithm from reccMonsters.js
Fixed bugs in algorithm.py and Monster.py so it runs successfully (for my very limited two test cases). I had to scrap some of the multiattack parsing because there are too many different ways to phrase a multiattack, so I put a temporary fix in and commented out the old code. I also set all player stats to 10 as another temporary fix. I accounted for some edge cases and null values when reading from our api as well. There might be some stray print statements left in algorithm and monster from my debugging--feel free to remove. My next goal is going to be for the algorithm to return the full json rather than just the monster name, so we can display the statblock on the front end